Refactor: move mocks.dart into subdir
