+++ test/pages/journal/journal_page_test.dart
@@ -20,7 +20,7 @@ import 'package:lotti/widgets/journal/entry_tools.dart';
 import 'package:material_design_icons_flutter/material_design_icons_flutter.dart';
 import 'package:mocktail/mocktail.dart';
 
-import '../../mocks.dart';
+import '../../mocks/mocks.dart';
 import '../../test_data.dart';
 import '../../widget_test_utils.dart';
