+++ test/pages/tasks/tasks_page_test.dart
@@ -17,7 +17,7 @@ import 'package:lotti/themes/themes_service.dart';
 import 'package:lotti/widgets/journal/entry_tools.dart';
 import 'package:mocktail/mocktail.dart';
 
-import '../../mocks.dart';
+import '../../mocks/mocks.dart';
 import '../../test_data.dart';
 import '../../widget_test_utils.dart';
