+++ test/widgets/charts/dashboard_measurables_test.dart
@@ -12,7 +12,7 @@ import 'package:lotti/themes/themes_service.dart';
 import 'package:lotti/widgets/charts/dashboard_measurables_chart.dart';
 import 'package:mocktail/mocktail.dart';
 
-import '../../mocks.dart';
+import '../../mocks/mocks.dart';
 import '../../test_data.dart';
 import '../../widget_test_utils.dart';
