+++ test/widgets/sync/imap_config_actions_test.dart
@@ -8,7 +8,7 @@ import 'package:lotti/themes/themes_service.dart';
 import 'package:lotti/widgets/sync/imap_config_actions.dart';
 import 'package:mocktail/mocktail.dart';
 
-import '../../mocks.dart';
+import '../../mocks/mocks.dart';
 import '../../mocks/sync_config_test_mocks.dart';
 import '../../test_data/sync_config_test_data.dart';
 import '../../widget_test_utils.dart';
